Rearrange files into mock.go and interface.go
